+++ main_test.go
@@ -7,7 +7,6 @@ import (
 	"time"
 
 	"github.com/labstack/echo"
-	"github.com/philips-software/logproxy/shared"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -99,15 +98,4 @@ func TestMissingKeys(t *testing.T) {
 	os.Setenv("HSDP_LOGINGESTOR_KEY", "")
 
 	assert.Equal(t, 20, realMain(echoChan))
-}
-
-func TestPluginManager(t *testing.T) {
-	pluginManager := &shared.PluginManager{}
-	err := pluginManager.Discover()
-	assert.Nil(t, err)
-
-	err = pluginManager.LoadAll()
-	assert.Nil(t, err)
-
-	assert.Equal(t, 0, len(pluginManager.Plugins()))
 }
